fix: fix ordering and pnpm lock lol
