--- components/module/index.tsx
@@ -1,3 +1,0 @@
-export * from './about';
-export * from './blog';
-export * from './experience';
